requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6

django-cors-headers==4.3.1
google-api-core==2.17.0
google-api-python-client==2.117.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gspread==6.0.1
httplib2==0.22.0
idna==3.6
lxml==5.1.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
pandas==2.2.0
pillow==10.2.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2024.1
pywin32==306
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
StrEnum==0.4.15
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
